+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/UncheckedServletException.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc;
 
 public class UncheckedServletException extends RuntimeException {
