+++ mvc/src/main/java/com/interface21/webmvc/servlet/view/JspView.java
@@ -21,8 +21,11 @@ public class JspView implements View {
     }
 
     @Override
-    public void render(final Map<String, ?> model, final HttpServletRequest request, final HttpServletResponse response)
-            throws Exception {
+    public void render(
+            final Map<String, ?> model,
+            final HttpServletRequest request,
+            final HttpServletResponse response
+    ) throws Exception {
         if (viewName.startsWith(JspView.REDIRECT_PREFIX)) {
             response.sendRedirect(viewName.substring(JspView.REDIRECT_PREFIX.length()));
             return;
